* Fixed Bug 104637
Functions are registeren in BuiltIn Registry rather than
createing a new one. Makes it compatible with Spark 1.5.2 in
HDP 2.3.4.x.

Test is done nightly, since we do not have an HDP 2.3.4 cluster
on push validation. However, tests with hive context are included
in this package as well (all tests are executed with SAP SqlContext
and with the HiveContext!)

Change-Id: I878183b868e0646bcc8d103b4cc0b609f57bb3cb